stocks/requirements.txt
appdirs==1.4.4
APScheduler==3.9.1
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.6
cycler==0.11.0
Django==4.1.2
exceptiongroup==1.0.0rc9
fonttools==4.38.0
greenlet==2.0.0
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.1
multitasking==0.0.11
numpy==1.23.4
outcome==1.2.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
sqlparse==0.4.3
termcolor==2.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0
yfinance==0.1.84
gunicorn==20.1.0